Change mapper classes location
