+++ app/src/main/kotlin/com/jpaya/englishisfun/abbreviations/mapper/AbbreviationMapper.kt
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2020 Jose Maria Payá Castillo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.jpaya.englishisfun.abbreviations.mapper
+
+import com.jpaya.englishisfun.abbreviations.data.db.RoomAbbreviationsItem
+import com.jpaya.englishisfun.abbreviations.data.network.model.AbbreviationsContent
+import com.jpaya.englishisfun.abbreviations.domain.Abbreviations
+import com.jpaya.englishisfun.abbreviations.ui.AbbreviationsListPresenter.AbbreviationsItem
+
+fun Abbreviations.toRoomItem() = RoomAbbreviationsItem(
+    id = id,
+    abbr = abbr,
+    desc = desc
+)
+
+fun RoomAbbreviationsItem.toDomain() = Abbreviations(
+    id = id,
+    abbr = abbr,
+    desc = desc
+)
+
+fun AbbreviationsContent.toDomain() = Abbreviations(
+    id = id,
+    abbr = abbr,
+    desc = desc
+)
+
+fun Abbreviations.toPresentation() = AbbreviationsItem(
+    id = id,
+    abbr = abbr,
+    desc = desc
+)
