+++ app/src/main/kotlin/com/jpaya/englishisfun/irregulars/data/network/NetworkDataSource.kt
@@ -17,7 +17,7 @@
 package com.jpaya.englishisfun.irregulars.data.network
 
 import com.jpaya.englishisfun.firestore.FireStoreClient
-import com.jpaya.englishisfun.irregulars.data.db.toDomain
+import com.jpaya.englishisfun.irregulars.mapper.toDomain
 import com.jpaya.englishisfun.irregulars.domain.Irregulars
 import javax.inject.Inject
 import javax.inject.Singleton
